Restructure to directories
