+++ packages/djangoflow_app/djangoflow_app_example/lib/widgets/app_version.dart
